refactor and extract model creator methods to helper
